+++ Demo/Demo/Views/GroupsList/GroupsListView.swift
@@ -7,34 +7,6 @@
 //
 
 import SwiftUI
-import Uploadcare
-
-class GroupsListStore: ObservableObject {
-	@Published var groups: [GroupViewData] = []
-	private var list: GroupsList?
-	var uploadcare: Uploadcare? {
-		didSet {
-			self.list = uploadcare?.listOfGroups()
-		}
-	}
-	
-	init(groups: [GroupViewData]) {
-        self.groups = groups
-    }
-	
-	func load(_ completionHandler: @escaping (GroupsList?, RESTAPIError?) -> Void) {
-		let query = GroupsListQuery()
-			.limit(5)
-			.ordering(.datetimeCreatedDESC)
-		
-		self.list?.get(withQuery: query, completionHandler)
-	}
-	
-	func loadNext(_ completionHandler: @escaping (GroupsList?, RESTAPIError?) -> Void) {
-		self.list?.nextPage(completionHandler)
-	}
-}
-
 
 struct GroupsListView: View {
 	@EnvironmentObject var api: APIStore
